EthSwt_SWIC_Timeタイマー32ビット化＋ラップアラウンドの処理　MCSESW-4150

+++ src/PE3VM3/App/Central/esw/EthSwt/Core/EthSwt_SWIC_Time.h
@@ -8,7 +8,7 @@
 /* -------------------------------------------------------------------------- */
 void EthSwt_SWIC_Time_Init(void);
 void EthSwt_SWIC_Time_HiProc(void);
-uint16 EthSwt_SWIC_Time_Get(void);
+uint32 EthSwt_SWIC_Time_Get(void);
 /* -------------------------------------------------------------------------- */
 #endif /* ETHSWT_SWIC_TIME_H */
 /* -------------------------------------------------------------------------- */
